+++ CHANGELOG.md
@@ -1,4 +1,10 @@
-## Version 1.1
+## Version 1.2 - 2023.12.xx
+
+* New commands for fixing all violations or organizing all imports in your workspace
+* Support (and prefer) `ruff format` in addition to `black`
+
+
+## Version 1.1 - 2023.12.14
 
 * Task, command, and settings for cleaning up Python cache files and build artifacts (#4)
 * New setting for Pip's `--upgrade-strategy` option (#3)